FEATURE: Authentication in the core server
This initial checkin plumbs through just enough of the core server
authentication to allow you to login via the web tier, and use that
login to control access to web tier controllers and re-use the
authentication for model updates against the core server.

QA: You should be able to log in now. :-)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>