BUG: Change back from relative paths to absolute paths for images in stylesheets.
While I'd love to have relative URLs, this interferes with Rails'
caching support for concatenating/caching stylesheets. What used to be a
stylesheet at the URL /stylesheets/foo is now a stylesheet at
/stylesheets/cache/foo, which breaks our relative URLs.

The most robust way to fix this is to just not use relative URLs, sadly.

QA: Make sure all the images throughout the site still load.

Reviewed-by: Pete Stuart <pete.stuart@blist.com>