BUG: Redo session support in rails/ web/ and core/ to use a unified cookie
This checkin changes the way we handle authentication across all three
servers currently in use. In order to prevent weird issues with the
cookies getting out-of-sync, we now use a modified form of the rails
deafult cookie:

Cookie       := Core_Data::Rails_Cookie
Core_Data    := User_ID Expiration Salt Core_Signature
Rails_Cookie := Marshalled_Data--Rails_Signature

Accomplishing this meant implementing some or all of the changes across
all three servers - you'll have to rebuild your core server to make this
work, and clear any existing cookies on your development machines.

One nice advantage of the new implementation for web/ is that it's
implemented in Rack - meaning that if we ever want to replace Rails or
use a smaller framework like Sinatra or Rails Metal, we still get access
to the session information.

QA: Test login, logout, and session expiration scenarios. Sessions
expire after 1 hour of inactivity, and should automatically extend your
hour as you use the application. A good thing to test would be to use
the app, change your system clock to be 2 hours later, and continue
using the application. Remember me functionality is not part of this
checkin, and it's expected that it doesn't do anything yet.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>