BUG: Yet more session bugs in alpha.blist.com
My prior commit changed the wrong spot - the swf_controller in the old
rails project isn't used anymore for setting variables that get passed
to the swf; web/app/controllers/swf_controller.rb is!

In the newer swf_controller, I'm not sure how this was ever working - we
passed in session.id which is technically a Fixnum, not the full session
string. I changed it to read the full session string from the cookies,
and split out only the rails portion of the string.

QA: Once again, accessing private blists on alpha.blist.com should work.

Reviewed-by: Justin Friedl <justin.friedl@blist.com>