FEATURE: Ability to submit your profile photo as part of the account creation process
This adds the ability to submit your profile photo with the rest of your
details when you sign up. Under the covers, it's a three-step process
where we create your account, log you in, then upload your profile photo
to the core server using your new account credentials.

* Added the multipart-post gem to assist with Net::HTTPs ineptitude in
  submitting multipart forms on its own.
* Refactored the config.gem stuff in environment.rb so that some gems
  are declared before updating our load path (ie: ones that aren't
  vendored, because they require native code) and some that happen after
  tweaking our load path.
* Plumbed through a generic multipart_post_file method on the Model, and
  added profile_image= accessor on the User class.
* Renamed profile_image to profile_image_path to avoid confusion with
  profile_image=(file) which expects a File object.
* Made the signup form an actual multipart form, so we get the content
  of the file rather than the filename. :-)

QA: You should check that you can upload a profile picture as part of
the signup flow, change the profile picture later as part of the account
maintenance/profile page, and verify that the profile picture continues
to show up everywhere it should.

There are some edge cases worth invtigating, too - as a starting point,
uploading a non-image should still create the account and log you in,
just not give you a profile picture. I'd also try creatively renaming
the image files and extensions to try to break it.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>