FEATURE: Giant performance speedup for /blists
This is a collection of mostly-small changes to various parts of our
codebase that were identified as slow when profiling a GET /blists
request with 300 views. Some highlights
* I reduced the number of times we call method_missing by dynamically
  generating (simpler) functions that are called on subsequent calls.
* I removed a lot of references to rails helpers in the _blists_row
  partial. While the helpers are generally okay, they break down when
  you're rendering the same partial 300 times with slightly different
  output on each. FormTagHelper, link_to, image_submit_tag, etc. are all
  gone from this partial.
* I used ActiveSupport::Memoizable to Memoize a few frequently-asked for
  properties.
* I sped up humane_date by not turing a time to a string then back to a
  time each time we called it, because the ruby code to parse strings
  into Time objects is slow.
* I rewrote menu_tag helper to use a StringIO instead of reallocating
  strings all over the place.

QA: Test all the functionality on /blists again, and in particular let
us know how things fare with accessing /blists in an account with a lot
(200+) blists. It should be noticeably faster, but there's still room to
improve - we're looking at about 2 secs to download it instead of 10-15.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>