FEATURE: Forgot password support
This checkin (re-)introduces the ability to reset a forgotten password
by emailing a special token to the user. The logic for generating the
reset codes and emailing the user is part of the core server; the
frontend merely provides some simpler forms for entering the user's
login or email, and another form to reset the password.

Unlike the prior implementation, I've moved the password resets out into
a separate table from the users table - the data is otherwise sparse,
and it's also nice to have a history of reset requests, as well as
keeping track of when the requests occurred. Until the old rails code is
deprecated, I couldn't get rid of the existing code, however.

QA: The workflow for this should be pretty self explanatory -
1. From the login dialog, click "Forgot your password / username?"
2. Enter either your username OR email address to generate a reset code,
which will be emailed to the address on file for that account.
3. Click the link in your email to load a page asking for a new password
4. After picking a new password, you're redirected to the homepage.

Interesting edge cases worth testing:
* Only the last generated reset code should be valid for a user. If you
generate more than one reset code for a user, only the last one will
allow you to reset a user's password.
* I made a number of refactorings to the existing code for setting your
password. It's worth re-testing the signup process and the ability to
change your password as a logged-in user, rather than the "can't login"
method I worked on.

Reviewed-by: Kostub Deshmukh <kostub.deshmukh@blist.com>