FEATURE: Friendly 404/500 error pages, initial work on static content.
This checkin introduces new routes and basic templates for some static
content on the site - the beginnings of the about page, as well as
better-looking 404 (Page not found) and 500 (Server error) pages.

If you need to test changes to this locally, you'll want to uncomment a
line in application_controller.rb:
  alias_method :rescue_action_locally, :rescue_action_in_public
Doing so makes it so that you no longer get call stacks on crashes when
accessing your server locally; instead you get what the public sees.

QA: To test the 404 page, simply enter a bogus URL. To test a 500 server
error page... you'll have to figure out some way to make the server
crash. It looks a lot like the 404 page, but with different text.

The static content so far is a basic copy of what's on socrata.com right
now - you can visit /about and /about/government_datasets to see the
content; at this point I'm only concerned with the fact that it works,
not whether the content is useful or whether it looks good. (Jon Byrum
will eventually own and update this content)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>