BUG: Fix for file imports, profile picture uploads
The way were currently handle imports is extremely fragile - the AJAX
upload control relies on an IFRAME for the upload, which means that we
don't directly get the response payload back like we would via
XMLHttpRequests. The net result is that it's usually wrapped with a
<pre> tag, but not always the same - Google Chrome likes to add
attributes to the <pre> tag, for instance.

This checkin changes the import code to be more resilient to handling
odd payloads:
* We strip off the <pre> tag from a plain-text response automagically
* Imports do less ugly string parsing for error cases
* Profile pictures automagically make the response a JSON payload
* Because we upload via an AJAX upload IFRAME, and Firefox doesn't
  support application/json as a MIME type, we've hard-coded the
  content-type of the Imports, ProfileImages, and FilesServices to
  use text/plain.
* I fixed a bug in uploading profile pictures with IE, where we
  cache-busted using Date.now() which isn't available cross-browser.

There's one major side-effect to this: IE is broken for file uploads
that fail (uploads that succeed are fine; failures are bad.)

QA: Test file imports, changing your profile photo. (IE & Google Chrome
at the minimum)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>