BUG 675185: Cleanup error handling for file imports, fixes for IE
Now that imports always returns a text/plain payload that is a JSON
object, we can get rid of some really ugly HTML parsing code in the
import javascript, and rely on some well-formed JSON instead.

This also fixes a longstanding, impossible-to-track down bug with IE not
setting errors correctly. It wasn't possible to access the IFRAME's
contentDocument when the upload failed - after much soul searching, I
realized that it's because our JSON error payload is < 512 bytes, which
causes IE to replace the error payload with its "friendly" error
messages. That friendly message comes from a resource DLL in IE, with a
URL like res://ieframe.dll/http_500.htm - this isn't the same domain as
our content, so you can't access it. BOO!! HISS!! LAME!!

QA: If you import something that won't work (for example, take an image
file and rename it to have a .csv extension, then import it) you should
see an error message shortly after starting the import, informing you of
the failure. The spinner will also go away.

Be sure to test this cross-browser, and when using IE turn on and off
the "Friendly error messages" feature:
http://prideafrica.blogspot.com/2007/01/ie-friendly-http-error-messages-not.html

Reviewed-by: Pete Stuart <pete.stuart@blist.com>
Reviewed-by: Kostub Deshmukh <kostub.deshmukh@blist.com>