NOOP: Static content in a submodule
This checkin adds a new git submodule for maintaining the static content
throughout the socrata.com website. It's deployed separately from the
main content, in an effort to make it easier for us to redeploy it
independently from the engineering shipping cycles.

To use this, you'll have to run "git submodule update --init" to start
seeing the static content.

Notable changes:
* Added routing for all known top-level content
* Refactored static logic into a ruby module that's mixed in to the
  various top-level controllers for different static content.
* Added a git submodule for static-content
* Removed 'about' views and replaced with a symlink to appropriate
  static-content section.

QA: Not ready for testing yet - most of the content isn't the final
styling yet, or not in place.

Reviewed-by: Chris Metcalf <chris.metcalf@blist.com>
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>