BUG: Rewrote column resizing.
Column resizing is now handled properly via an event on the model.  Decoupled column sizing from column rendering so that
we only need sizing logic implemented in one place vs. the old model where we sized once on render and again when the user
sized a column.  Rewrote pointer management in the grid.  We now do it programmatically rather than via CSS.  This allows
us to change the cursor e.g. when the cursor hovers over the area on a column that allows for resizing.  The old method
required us to move little divs around which isn't viable w/ the new nested table support.  Used all of this to fix long
standing sizing bugs and to implement nested table (a.k.a BnB) sizing.

QA: Ensure that you can properly size top-level columns and blist-in-blist columns in the JS (read-only) grid.

Reviewed-by: jeff.scherpelz@blist.com