FEATURE: Contact Us form
When running in development mode, the contact form sends an email to
engineering@socrata.com. When running in production mode, the contact
form sends an email to feedback@socrata.com. There's an additional
Sender header so it shows up as "on behalf of" us in some mail clients,
including Outlook. It also logs the requestor's IP address in a custom
header X-Contact-IP, if you view the raw headers of the message. Once
the form is submitted, the form is replaced with a simple "Thank you"
message.

QA: Test that you can recieve emails on both test
(http://test.socrata.com) and production (http://alpha.socrata.com) as
well as the links throughout the site that should lead to the contact us
form do so.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@blist.com>