FEATURE: Change *.socrata.com to use the "blist" skin, not the "gov" one.
Now that the blist skin has changed to be mostly like the gov skin, we
want to jump back to using the blist skin, since that has the
appropriate header/footer content, logos, etc.

QA: loading test.socrata.com and alpha.socrata.com should now show the
blist logo (soon to be replaced) and not the fake data.gov logo, along
with the socrata-specific footer treatment instead of the
government-themed one.

Reviewed-by: Chris Metcalf <chris.metcalf@blist.com>