FEATURE: Show the dataset name in the page title, and add meta description + keywords
* Added a meta_tags helper
* Added logic in BlistsController#show to display meta tags when viewing
  a dataset.

QA: When viewing a dataset, you should see the title of the page now has
the title of the dataset as the first component. Additionally, if you
view the HTML, you will see two new <meta> tags in the HTML <head>
block:

<meta name="description" value="..." />
<meta name="keywords" value="..." />

Description will always show up, and could be any of the following:
* The long description of the dataset, if the user entered one.
* The text "View this dataset, last updated _____" if they don't have a
  description but we do have metadata about when the rows were last
  updated (true of all but the oldest untouched datasets)
* The text "View this dataset" as a fallback.

Keywords only show up if the user entered tags, and shows up as a
comma-separated list of tags.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>