BUG: Show the "blist is now Socrata" popup when the user came from www.blist.com
The 301 redirect simply passes through the referrer of the original request. If you didn't click through to www.blist.com from a blist.com domain, you wouldn't see the proper popup.

I've added a ref= query parameter we can set in the redirect that will force the proper pop-up to show.

QA:
Visit the following URL and confirm that the "blist is now Socrata" modal is shown:
http://test.socrata.com/?ref=www.blist.com

Reviewed-by: Paul Paradise <paul.paradise@blist.com>