FEATURE: Remember me
This introduces a 2nd authentication cookie, which unlike the
_blist_session_id is saved across browser restarts. By default, remember
me lasts for two weeks. The token itself is a 32-character random
alphanumeric string, which is stored in a database table along with an
expiration date and user to authenticate as.

Both the new rails and core server support authenticating via the
remember me token; the old rails code won't do it, but luckily for us
you first always hit a host page that's controlled by the new rails
process, so you shouldn't end up in a situation where you're not authed
and should be.

QA: Heavily test login/logout scenarios, using both the remember me
checkbox and not using it. It's important to remember that the remember
me is most useful across browser restarts - you should be able to quit
your browser, start it up again, and still be logged in when you visit
again, regardless of what page you visit first on restart. You should
also be able to remember your session on multiple browsers independently
- each has its own remember token, and logging out from Socrata removes
  the remember token from that individual instance.

Reviewed-by: Kostub Deshmukh <kostub.deshmukh@socrata.com>