NOOP: Remove the _blist_session_id from old SWF
If you use Firebug and look at the HTTP requests that our SWF makes,
you'll notice that pretty much every URL request has
"?_blist_session_id=undefined" added to it. This *USED* to be a
mechanism by which we forcibly set the session ID because the SWF didn't
always send the session cookie - if memory serves correctly, it's
largely for file uploads.

Clearly, with the session=undefined, we're not using it anymore. I opted
to get rid of it rather than try to fix it.

QA: Test the functionality that still relies on the old SWF: editing the
grid, adding columns, uploading documents and photos in a dataset, etc.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>