FEATURE: Remove /data route, make everything reference /
Static page caching was broken for the homepage because we expect the
homepage payload to have the text "<title>Socrata" in it - but we
renamed this inadvertently to "Data | Socrata | Making Data Social"

This all happened because we have another route for the DataController
(/data) that's essentially the same thing as the homepage. Google and
other search engines are likely penalizing us for having the duplicate
content, so I decided to get rid of the /data URL and just keep /.

QA: Ensure all the links to view the Data tab (or related filtered
content) still work. Request to /data should be automatically redirected
to / with an HTTP 301 (Moved Permanently) error - you can inspect this
using a tool like Firebug.

Reviewed-by: Pete Stuart <pete.stuart@socrata.com>