BUG: Searching on most pages other than the data tab was failing
A number of spots still search by having a form pointed at /data, and we
weren't passing the search parameter. I changed the spot I could find
that still used /data, as well as made it pass the parameters through as
part of the redirect, just in case there are other spots.

QA: Self-explanatory.

Reviewed-by: Pete Stuart <pete.stuart@socrata.com>