FEATURE: Added an auto-generated sitemap to help Google find user pages and datasets.
By default, the sitemap is cached indefinitely, so the cache must be cleared periodically (with web/script/clean_sitemap_cache.sh) in order to update it.


QA: /sitemap.xml should be a xml sitemap index file, and the sub-sitemaps it links to should be xml sitemaps containing links pointing to all the users and datasets.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>