BUG: Changing your profile picture in Chrome didn't refresh nicely
We were setting the src attribute of the img tag to "" while doing the
change; this caused issues because Chrome (among other browsers) would
display a broken image.

Now we fade out while we're waiting for the upload to complete, and then
replace it with the new image, preloaded in the background and nicely
faded in. The net effect if you have a decent connection is that the
image seamlessly fades in/out.

QA: Change your profile picture.

Reviewed-by: Pete Stuart <pete.stuart@socrata.com>