BUG: Sign Up - Profile pictures added donot show up
A weird interaction between Rails and an old version of Rack meant that
we had a buggy file-upload system that lost the original filename.
Change ed3b8f5c introduced validation on the image filename on the core
server, which promptly caused this to fail.

Upgrading to Rack 1.0.0 (from 0.9.1) allows us access to the
original_path of the file (rather than the random tempname generated on
the app server), and then all we have to do is plumb it through.

QA: http://www.pivotaltracker.com/story/show/827579

Reviewed-by: Kostub Deshmukh <kostub.deshmukh@socrata.com>