BUG: Requests to /sitemaps.xml were still being routed by Rails
Commit 503f602335cb90d2144b487efd2ea829d697d59e changed our sitemap code
to use Java instead of Rails, but forgot to pull out the routing from
Rails itself. Now visiting that page in Rails results in a an HTTP 500
error instead of and 404. This checkin corrects that; you'll still need
some Apache RewriteRules to end up with the real content.

QA: /sitemaps.xml shouldn't return an HTTP 500 error.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>

Conflicts:

	web/static-content