BUG: View With No Row Views Crashes Stats
QA:

Go to the dataset create screen and enter a name. Click "create" and
wait for the page to try to load the dataset page. Before it loads all
the way, close the tab (the timing it a bit weird). Then go to "My
Socrata" and find the link to the dataset that you just created. Copy
the link and paste it into the address bar, but add "/stats" to the end
of it.

Result:

The stats page should load no problems.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>