SPARTA: Consolidate gobs of YAML
This checkin consolidates a bunch of random little YAML file we've been
creating and scattering through the project into a single YAML file
named config.yml.

The one change this does introduce is that the test.socrata.com
deployment will no longer run in the "production" stage, which gets
confused with production stage in, well... production. It's now called
"staging". Instead of having two "levels" to our deployment terminology
(the webistrano stage and the rails env) we can start consolidatinig it
to just use the rails env.

QA: Nothing specific to test, but this affects how we load up and use
some environment-specific parameters like the Socrata RSS feed shown in
the "From the Blog" section, the connection info for the multiuser
service when using the old flash grid, and how to contact the core
server itself.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>