BUG: Fix dynamic popup login to use SSL
This was a pain to figure out, but the code is simple. :-) Our old
method of logging in via AJAX (used for posting comments and saving temp
views created when not logged in) wouldn't work with SSL. The
same-origin policy on browsers means that we can't make an
XMLHttpRequest to https://www.socrata.com from a page hosted on
http://www.socrata.com - they're not the same host.

We work around this by using JSONP to login instead. This writes out a
magic <script> tag into the page content, which forces the browser to
load the script's content just like it would any script, image, etc.
loaded on the page - bypassing the same-origin policy. The script
accepts a callback method, allowing it to call back into our
inline-login.js code to complete the authentication loop.

The really magical part is that the login cookies are set correctly.
Although the same-origin policy prevents us from making an AJAX request
to the HTTPS version of the site, cookies don't care about the protocol
or port, only the host. So when we login via JSONP, the response headers
include the appropriate Set-Cookie information, which happens to work
because the HTTPS content we retrieved is on the same host name as the
HTTP content. (In other words, this wouldn't work in most JSONP
scenarios, where the host varies instead of the port or protocol.)

QA: Please test login/logout functionality, particularly those scenarios
that involve logging in after creating some content as an
unauthenticated user, such as posting a comment to a dataset or saving a
temporary view.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>