BUG: Profile pages should return 404 for nonexistent profiles (#1 Rails crash from Hoptoad)
Until now, we weren't properly handling 404's from the core server
properly - they'd turn into a generic CoreServerError, which generally
turned into an HTTP 500 error in Rails unless we did something about it,
which we rarely ever did. In general, failing to get something from the
core server really does mean that the resource we're trying to load
doesn't exist, so we might as well show a 404 to the end user, too.

* Created a new CoreServer namespace, and started building up a list of
  errors modeled after ActiveResource. Consolidated the existing
  CoreServerError into the same namespace, but it needs more work to
  really start breaking these down.
* Setup rescue_from block in application_controller.rb to treat all
  CoreServer::ResourceNotFound errors as a generic 404 for us.
* Changed a spot in blists_controller.rb where we were catching a 404
  from the core server to use the new mechanism, and also change it to
  actually return a 404 instead of a 200.
* Updated all the spots where we were using CoreServerError to be a
  CoreServer::CoreServerError.

QA: You should see a "Page not found" error instead of a "Something's
wrong" error when visiting a URL for a user profile that doesn't exist;
change the abcd-1234 to something random to to test. Similarly, when
viewing a dataset page that doesn't exist (or was previously deleted)
you should get a custom error message indicating that - if you check
with Firebug or another tool, you should notice all these "not found"
pages now are HTTP 404 Not Found pages instead of an HTTP 200 OK.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>