BUG: Fix 18,000 broken links with one character! (/close)
Google Webmaster Tools was reporting we have about 18,998 broken links.
Most of them appear to be of the form:

http://www.socrata.com/dataset/Friendly-Name/xxxx-xxxx/close

It turns out we missed a hash mark on a link/button to close the
licensing and attribution section of the dataset's about page; this
still worked functionally, but non-Javascript browsers (like a web
crawler) would try to follow the link and get a 404.

QA: The you should be able to close the licensing and attribution
section of a dataset's about page, after you expand it to edit it.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>