BUG: Sign Up/In link appears for an already signed in user.
The repro steps in the bug gave me some hints as to what was wrong, but
in the end it was an interesting conglomeration of two different bugs:

1. We weren't properly respecting the "Remember me" setting on login -
we treated a not-checked box as true, because it's a string "0" which
evaluates to true. In other words, even when you told us not to remember
you, we did.
2. With remember me fixed, I noticed a problem where the core server was
mistakenly overwriting the _blist_session_id cookie with a max-age
parameter, which turns it from a cookie that expires at the end of the
browser session into a cookie that expires after an hour, regardless of
whether you exit the browser or not.

QA: http://www.pivotaltracker.com/story/show/1011050
You'll want to give login/logout and particularly remember me
functionality a good run-through, particularly around browser
exit/reload cycles. (Remember me => logins persist across browser
restarts)

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>