BUG: Login via popup was always setting remember_me token
Okay, this is really annoying. In HTML, checkboxes have a value, say
"1", that is sent as part of a form only when the checkbox is checked.
If it's not checked, nothing related to it is sent.

Because Rails wants to have an easy way to do a form based on an
ActiveRecord model and be able to call update_attributes!(params), but
if you wanted to unset a value on a model, if the param didn't exist, it
wouldn't know to update it. So Rails adds a secret, hidden checkbox with
the value set to "0" above the real checkbox. Rails always takes the
last value, so this works.

When logging in via jQuery, we were just selecting all the form inputs
($form.find(':input')) which included the unchecked checkbox - now all
the sudden we're submitting the form with a remember_me value set to
'1', because normally it wouldn't be submitted but jQuery selected it.

I couldn't find a good jQuery.selectFormFieldsYouWouldNormallySubmit(),
so I just selected the things I knew I'd need to make the login work.
Hopefully we won't get bitten by this again.

QA: http://www.pivotaltracker.com/story/show/1086132

Reviewed-by: Pete Stuart <pete.stuart@socrata.com>