FEATURE: Sharing Part II API changes on the server caused a few ripple effects on the client.
QA: The sharing API has changed. Verify that shares still work, the
shares are displayed correctly in the sharing tab and it is still
possible to share (albeit still difficult without the new UI) a
dataset. When you share a dataset, all views on the dataset should
automatically be shared.
Verify that sharing also works for views and displays additional
sharing information from the dataset.

Reviewed-by: Pete Stuart <pete.stuart@socrata.com>
Reviewed-by: Sam Gibson <sam.gibson@socrata.com>