SPARTA: Log request times to the core server
Up until this checkin, we've been making mostly-educated guesses about
how slow Rails is vs. the core server when creating an HTML page in
Rails. Since we often have to make multiple requests to the core server
(sometimes even while rendering the view), it's not always clear how
much those serial requests add up. Now we know.

Much like when you render a partial and see how long it takes to render
each partial in a request, you'll now see the path to a core server
request and the time the request took. Additionally, at the end of the
request in the summary, we include an aggregate of the number of
milliseconds spend in the core server for that request, and adjust the
view's render time to not count time spent waiting on the core server as
part of it. An example request, snipped to show the relevant parts:

  Processing HomesController#show (for 127.0.0.1 at 2009-08-17 17:09:42) [GET]
    Parameters: {"action"=>"show", "controller"=>"homes"}
  Core server request: /users/nw2t-u8ya.json (13.7ms)
  Request domain: localhost, Current locale: blist
  Core server request: /users/nw2t-u8ya/views.json (186.9ms)
  Core server request: /activities.json?inNetwork=true&maxResults=5 (12.3ms)
  Rendering template within layouts/main
  Rendering homes/show
  Rendered shared/_tabs (0.4ms)
  Core server request: /users/nw2t-u8ya/views.json (149.1ms)
  Core server request: /users/nw2t-u8ya/views.json (231.8ms)
  [Rendered a bunch more partials...]
  Completed in 1175ms (View: 32, Core Server: 594) | 200 OK [http://localhost/home]

In order to handle this, I did a few things:
* I split the logic for connecting to the core server and getting a
  response back out of the Model class. Now the Model is primarily used
  for parsing, accessor functions, etc.
* Following the model of ActiveRecord, I introduced a Connection class
  for the core server that keeps track of benchmark times for a series
  of requests, and provides a reset_runtime method to reset the counter
  back to zero and report the last counter.
* I overwrote the ActionController::Benchmarking module with our own
  CoreServerBenchmarking module. This is a bit of a hack because I wish
  Rails had better extensibility points for adding our own benchmark
  data to the log, but it really can't be done cleanly. This will likely
  have to be re-done whenever we upgrade Rails.

QA: Functionally, nothing changed. However, just about every single spot
in the frontend that makes a request to the core server (which is
essentially every page) had small changes made to it. I tried to make
sure loading everything worked well, but I'd pay particular attention to
form submissions and AJAX style updates since they got much less
testing.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>