BUG: Fix crash in CoreServer::Connection that masked real core server crashes
We had a number of crashes in Hoptoad of the form
  NoMethodError: undefined method `data' for #<Hash:0xdeadbeef>

This was a bug in the connection code refactor; we now need to access
the hash directly rather than the data property on the now-nonexistent
Model instance. Note that we were already crashing from something else
for these errors (the core server returned a non-success status code)
but we were masking what the issue was.

QA: I was able to repro this by trying to create a profile link without
selecting something from the dropdown for the link type, but from the
end-user perspective nothing really changes here, just how we report the
error internally.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>