BUG: Unable to change column properties, convert columns, change widget customizations, etc.
This checkin tries to fix an entire class of bug I introduced
inadvertently: calling the instance method parse() on a descendant of
the Model class instead of the class method parse(). The class method is
what we want (turning a JSON string into a model object), whereas the
instance method goes through method_missing and tries to get the 'parse'
property from the current instance - usually, that's nil.

To help prevent this from occurring in the future, the Model class now
defines an instance method parse() that tells you that you probably
wanted the class method. Eventually we should probably break the parsing
code out into it's own class entirely, much like we did with the
connection code.

QA: http://www.pivotaltracker.com/story/show/1116198
    http://www.pivotaltracker.com/story/show/1116196
    http://www.pivotaltracker.com/story/show/1116195
There are similar problems with saving SDP customizations and logging
opening a view to update the stats, but I didn't see any bugs on those.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>