FEATURE: Crash the Rails app when you write slow code
From this checkpoint forward, if you write code in the rails app that
makes too many requests of the core server in a single client request,
we'll raise an exception and crash. Additionally, a message will get
logged to Hoptoad indicating the slow request.

This is not true of production - we'll still get the notice in Hoptoad,
but the request will be served to the client.

As an additional bonus, the core server request logging now includes a
count of the total number of requests, so you don't have to count. :-)

QA: While functionally nothing changed, you're likely to see some pages
that were slow before no longer work at all now. (Usually, it'll look
like a "Something's wrong" page.) We get error messages indicating
they're a problem, but you should still file bugs against it.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>