FEATURE: Initial login via OpenID support
This checkin merely plumbs through the Javascript and links necessary to
pop up an RPX login dialog, and provides an endpoint we get redirected
back to without doing anything useful. You can try logging in with an
OpenID credential, but it won't actually do anything useful yet after
you enter your username and password.

QA: At this point you can try using the login box, but it's not wired up
end-to-end so it's not really worth looking at very hard. Don't file
bugs on this yet.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>