BUG: Fix broken instantiation of CoreServer::TooManyRequests exception
A last-minute change after a code review bit me - I added a 3rd argument
to the constructor without changing it elsewhere. The net effect is that
if the server was supposed to crash from too many requests, it'll still
crash, but with the wrong error getting reported to Hoptoad.

This just fixes it so we throw the request we're expecting, instead of
an ArgumentError. Note that while this has no perceivable change in
staging, it will result in crashes on production that we don't want - so
it should get cherry-picked if we deploy the current test candidate.

QA: Nothing to test

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>