BUG: Fix per-request local cache to work again
My earlier refactors for the CoreServer::Connection broke the rack
middleware that enabled a per-request cache. This change brings it back.
In addition, it alters our RequestStore to not bother calling super,
which merely logged a duplicate "cache read/write" line to our logs
since the caching infrastructure is designed as a two tier cache (local
and remote, typically memcache) of which we're only using the local
component.

QA: A lot of pages should feel faster.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>