FEATURE: Widget Customization — making gradients work, fixing minor bugs.
QA:
Alright, here's a big one.  Go to any dataset you own with an account that's enabled for premium sdp use.  Add /publish to the url, and you should get the widget publishing interface.  The changes you make are autosaved, but if you hit cancel changes it will revert you back to the way it was at the beginning of the current session.  Templates are shared within each organization, and the three of us are currently part of the same test organization.  The designer isn't meant to account for multiple people editing the same template at once, so make sure you don't do that or you'll end up overwriting each other's changes.  Note that the way we update the preview in the publishing interface is a completely different set of code from the way we actually render the widget for real, so even if one is broken the other is probably not.
The image upload stuff is a new set of code as well, so give it a good runthrough.  Images that you upload should be shared within your organization.  As with the other places where we do upload-type things, there's a known issue where sometimes when you click on the 'choose a file' button it won't give you the choose a file box but instead take you somewhere random.  This is because of the plugin we are using to handle those file uploads, and I'm going to work on replacing that plugin when I get back from India.  For now, you'll have to deal with it — a clue for you is that when your mouse cursor is a hand, you're going to go to the wrong place.  If it's the regular arrow, then you'll get the correct dialog.  It helps if your mouse enters the button from the top; don't ask me why.  There isn't any way to delete, rename, or replace images yet; that will come later.
There are small known issues with some of the preview settings that don't affect how the widget eventually renders for real — the show/hide row numbers and wrap header text options can make things look a bit weird for now.  Feel free to file a bug for those two things together (it's one root issue).  When you change the settings for allowing/disallowing public saved filters and the interstitial ("warn while leaving") it doesn't actually take effect in the player at the moment.  This isn't a super-critical issue since it's not an immediate visual change anyway, but feel free to file another bug for the two of them together.
This particular commit doesn't work in IE at all.  I'm hoping to have that solved before I leave today, but it might not happen.
Lastly, there are caching issues we have yet to solve with this whole customization thing.  Since we cache each variation of the entire widget on the server (these themes included) the first time anyone lands on them and don't let go of them for a very, very long time, you will probably not be able to see changes you make to the template on the actual widget once you've already loaded it.  You'll have to apply the same template to another dataset and load up that widget in order to see the changes you made to the theme.
The format for these custom themes is pretty similar to the whitehouse, black, and gov themes: each custom theme has a 4-4 id, so the final url of the widget ends up being http://www.socrata.com/widgets/4-4/4-4 for the custom themed widgets.
Also note that the list of "players using this template" is currently a fake set of data — we're still trying to figure out how best to store that data.
There is also currently no validation on the size of the player, etc.  That will be done eventually.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>
Reviewed-by: Paul Paradise <paul.paradise@socrata.com>